requirements.txt
Django==1.8.4
djangorestframework==3.2.2
ipython==4.0.0
